refactor AEChannelInfo and AEAudioFormat enums
put in a separate file for exposure to add-ons
